bpv7/cbr: fix macOS clang sign-compare build error in cbr_handleCrs
sdr_list_length() returns size_t (unsigned); the CRS-history ring-limit
check cast crsHistoryMax to (long), so clang rejected the unsigned vs
signed comparison under -Werror,-Wsign-compare and failed the macOS
build (gcc on Linux compiled it). crsHistoryMax is unsigned int, so cast
it to (size_t) to match the list length.

+++ bpv7/library/cbr.c
@@ -3175,7 +3175,7 @@ int	cbr_handleCrs(Sdr sdr, unsigned char *adminRecord, int length,
 				}
 				else if (cbrDb.crsHistoryMax > 0
 					&& sdr_list_length(sdr, cbrDb.crsHistory)
-						> (long) cbrDb.crsHistoryMax)
+						> (size_t) cbrDb.crsHistoryMax)
 				{
 					/*	Evict oldest entry.	*/
 					firstElt = sdr_list_first(sdr,
